+++ crates/sncast/tests/e2e/script/init.rs
@@ -16,7 +16,7 @@ fn test_script_init_happy_case() {
         .args(["script", "init", script_name]);
 
     snapbox.assert().stdout_matches(formatdoc! {r"
-        Warning: [..]
+        [WARNING] [..]
         command: script init
         message: Successfully initialized `{script_name}` at [..]/scripts/{script_name}
     "});
@@ -124,7 +124,7 @@ fn test_initialized_script_compiles() {
         .args(["script", "init", script_name]);
 
     snapbox.assert().stdout_matches(formatdoc! {r"
-        Warning: [..]
+        [WARNING] [..]
         command: script init
         message: Successfully initialized `{script_name}` at [..]/scripts/{script_name}
     "});